+++ _benchmarks/iris-mvc/main.go
@@ -8,5 +8,8 @@ import (
 func main() {
 	app := iris.New()
 	app.Controller("/api/values/{id}", new(controllers.ValuesController))
-	app.Run(iris.Addr(":5000"))
+
+	// 24 August 2017: Iris has a built'n version updater but we don't need it
+	// when benchmarking...
+	app.Run(iris.Addr(":5000"), iris.WithoutVersionChecker)
 }