+++ context/route.go
@@ -21,6 +21,12 @@ type RouteReadOnly interface {
 	// IsOnline returns true if the route is marked as "online" (state).
 	IsOnline() bool
 
+	// StaticPath returns the static part of the original, registered route path.
+	// if /user/{id} it will return /user
+	// if /user/{id}/friend/{friendid:int} it will return /user too
+	// if /assets/{filepath:path} it will return /assets.
+	StaticPath() string
+
 	// ResolvePath returns the formatted path's %v replaced with the args.
 	ResolvePath(args ...string) string
 }